Changed tests that were using false as a string.
There were tests to check that overloading actually overloads variables - and that loading
does not. Those tests used passing 'false' as a string, which is truthy. That behaviour is
to be changed now so 'true' and 'false' were replaced by 'set_by_environment' and
'set_by_file'.

 only when necessairy relied on 'false'
being passed as a string, which is truthy.